removing support username as it cannot change

+++ toolbox-kvm-openstack.sh
@@ -22,7 +22,9 @@ source /tmp/project_config.sh
 if [ ! -f "/tmp/linux.iso" ]; then
   pwd=`pwd`
   git clone https://github.com/mephmanx/centos-8-minimal.git /tmp/centos-8-minimal
-  curl -o /tmp/CentOS-Stream.iso $CENTOS_BASE -L
+  if [ ! -f "/tmp/CentOS-Stream.iso" ]; then
+    curl -o /tmp/CentOS-Stream.iso $CENTOS_BASE -L
+  fi
   cd /tmp/centos-8-minimal
   ./create_iso_in_container.sh "/tmp/CentOS-Stream.iso"
   mv /tmp/centos-8-minimal/CentOS-x86_64-minimal.iso /tmp/linux.iso
